requirements.txt
# Core scientific packages
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization and plotting
matplotlib>=3.4.0
seaborn>=0.11.0
pycirclize>=0.1.0  # For circular plots

# Bioinformatics
biopython>=1.79

# Machine Learning
scikit-learn>=1.0.0

# Utilities
tqdm>=4.62.0
PyYAML>=6.0
dataclasses-json>=0.5.7
colorlog>=6.6.0  # Optional for better logging